requirements.txt
streamlit
pandas
plotly
langchain-google-genai
langchain-core
google-generativeai
langchain-openai
python-dotenv
numpy
openpyxl
joblib
scikit-learn
transformers 
datasets 
torch
accelerate>=0.26.0
